Extract QuickBlueDarwin for iOS/macOS
