+++ packages/quick_blue/ios/Classes/QuickBlueDarwin.swift
@@ -1,0 +1,1 @@
+../../darwin/QuickBlueDarwin.swift
