+++ packages/quick_blue/macos/Classes/QuickBlueDarwin.swift
@@ -1,0 +1,1 @@
+../../darwin/QuickBlueDarwin.swift
